OvmfPkg/RiscVVirt: Modify the secure UART base address
In QEMU, there will be a new UART, the base address
of this UART is 0x10002000
Use this UART as the dedicate log out for StandaloneMM firmware.
This is only for the OpenSBI secure monitor based implementation.

The CoVE based implementation has no dedicated secure uart in HS mode
,it use the SBI DBCN ecall as default to passthrough the debug log out
to secure manager to output.

Signed-off-by: Yong Li <yong.li@intel.com>